fatfs: use correct return types in ioctl calls
GET_SECTOR_SIZE should return WORD (unsigned short) but returned
uint32_t.

Reference: http://elm-chan.org/fsw/ff/doc/dioctl.html

Closes https://github.com/espressif/esp-idf/pull/1031.

+++ components/fatfs/src/diskio_sdmmc.c
@@ -64,10 +64,10 @@ DRESULT ff_sdmmc_ioctl (BYTE pdrv, BYTE cmd, void* buff)
         case CTRL_SYNC:
             return RES_OK;
         case GET_SECTOR_COUNT:
-            *((uint32_t*) buff) = card->csd.capacity;
+            *((DWORD*) buff) = card->csd.capacity;
             return RES_OK;
         case GET_SECTOR_SIZE:
-            *((uint32_t*) buff) = card->csd.sector_size;
+            *((WORD*) buff) = card->csd.sector_size;
             return RES_OK;
         case GET_BLOCK_SIZE:
             return RES_ERROR;
